requirements.txt
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.2
murmurhash==1.0.9
numpy==1.24.3
packaging==23.1
pathy==0.10.1
preshed==3.0.8
pydantic==1.10.7
python-docx==0.8.11
requests==2.28.2
smart-open==6.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.9
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
